Address John's and Adam's comments.
