+++ sky/engine/core/script/dart_init.cc
@@ -118,10 +118,6 @@ static const char* kDartTraceStartupArgs[]{
     "--timeline_recorder=endless",
 };
 
-static const char* kVmCompleteTimelineArgs[]{
-    "--complete_timeline",
-};
-
 const char kFileUriPrefix[] = "file://";
 
 const char kDartFlags[] = "dart-flags";
@@ -517,9 +513,6 @@ void InitDartVM() {
 
   if (SkySettings::Get().trace_startup)
     args.append(kDartTraceStartupArgs, arraysize(kDartTraceStartupArgs));
-
-  if (SkySettings::Get().vm_complete_timeline)
-    args.append(kVmCompleteTimelineArgs, arraysize(kVmCompleteTimelineArgs));
 
   Vector<std::string> dart_flags;
   if (base::CommandLine::ForCurrentProcess()->HasSwitch(kDartFlags)) {
