+++ sky/engine/public/platform/sky_settings.h
@@ -19,7 +19,6 @@ struct SkySettings {
   bool start_paused = false;
   bool enable_dart_checked_mode = false;
   bool trace_startup = false;
-  bool vm_complete_timeline = false;
   std::string aot_snapshot_path;
 
   static const SkySettings& Get();
