+++ sky/shell/platform/mojo/BUILD.gn
@@ -13,8 +13,10 @@ mojo_native_application("mojo") {
     "application_impl.h",
     "content_handler_impl.cc",
     "content_handler_impl.h",
-    "dart_tracing.cc",
-    "dart_tracing.h",
+    "dart_trace_provider.cc",
+    "dart_trace_provider.h",
+    "dart_tracing_impl.cc",
+    "dart_tracing_impl.h",
     "main_mojo.cc",
     "platform_view_mojo.cc",
     "platform_view_mojo.h",
