+++ sky/shell/platform/mojo/main_mojo.cc
@@ -16,14 +16,13 @@
 #include "sky/engine/public/platform/sky_settings.h"
 #include "sky/shell/shell.h"
 #include "sky/shell/platform/mojo/content_handler_impl.h"
-#include "sky/shell/platform/mojo/dart_tracing.h"
+#include "sky/shell/platform/mojo/dart_tracing_impl.h"
 
 namespace sky {
 namespace shell {
 namespace {
 
 const char kEnableCheckedMode[] = "--enable-checked-mode";
-const char kVmCompleteTimeline[] = "--vm-complete-timeline";
 
 }  // namespace
 
@@ -45,7 +44,6 @@ class MojoApp : public mojo::ApplicationDelegate,
     blink::SkySettings settings;
     settings.enable_observatory = true;
     settings.enable_dart_checked_mode = app->HasArg(kEnableCheckedMode);
-    settings.vm_complete_timeline = app->HasArg(kVmCompleteTimeline);
     blink::SkySettings::Set(settings);
 
     Shell::Init();
@@ -64,7 +62,7 @@ class MojoApp : public mojo::ApplicationDelegate,
   }
 
   mojo::TracingImpl tracing_;
-  dart::DartTracingImpl dart_tracing_;
+  DartTracingImpl dart_tracing_;
 
   DISALLOW_COPY_AND_ASSIGN(MojoApp);
 };
